# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RECALL API
  version: 0.1.0
  description: 个性化游戏生成与分享平台 MVP API。
servers:
  - url: http://127.0.0.1:8080
    description: Local development
paths:
...
  /api/v1/admin/ai-settings:
    get:
      operationId: getAdminAISettings
      summary: 获取当前生效的 AI 配置
      description: API Key 永不回显；只返回是否配置、来源和脱敏末四位。
      tags: [Admin AI Settings]
      responses:
        "200":
          description: 当前 AI 配置及版本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAISettingsResponse"
        "401":
          $ref: "#/components/responses/APIError"
    put:
      operationId: publishAdminAISettings
      summary: 校验并发布一个新的 AI 配置版本
      tags: [Admin AI Settings]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminAISettingsRequest"
      responses:
        "200":
          description: 新发布的 AI 配置版本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAISettingsResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
  /api/v1/admin/ai-settings/test:
    post:
      operationId: testAdminAIConnection
      summary: 使用尚未保存的表单配置测试指定 AI 能力
      tags: [Admin AI Settings]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestAdminAIConnectionRequest"
      responses:
        "200":
          description: 服务真实请求测试成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestAdminAIConnectionResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "400":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "502":
          $ref: "#/components/responses/APIError"
...
  /api/v1/ai/love-letter/polish:
    post:
      operationId: polishLoveLetter
      summary: 润色情书；未配置文本模型时保留原文
      description: Key 为空时不调用外部模型，返回裁剪后的原文并将 skipped 设为 true。配置模型后，超时或供应商异常仍返回错误。
      tags: [AI Assistance]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoveLetterPolishRequest"
      responses:
        "200":
          description: 润色结果或无 Key 降级后的原文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoveLetterPolishResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
        "502":
          $ref: "#/components/responses/APIError"
        "504":
          $ref: "#/components/responses/APIError"
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
      example: http://127.0.0.1:5173
...
  responses:
    APIError:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    TextAISettings:
      type: object
      additionalProperties: false
      required: [enabled, provider, baseUrl, model, timeout, maxOutputTokens]
      properties:
        enabled:
          type: boolean
        provider:
          type: string
        baseUrl:
          type: string
          description: 能力启用时必须是有效服务地址；禁用时允许为空字符串。
        model:
          type: string
          maxLength: 256
        timeout:
          type: string
          examples: [30s]
        maxOutputTokens:
          type: integer
          minimum: 1
          maximum: 100000
    ImageModerationAISettings:
      allOf:
        - $ref: "#/components/schemas/TextAISettings"
    ImageToImageAISettings:
      type: object
      additionalProperties: false
      required: [enabled, provider, baseUrl, model, quality, timeout, maxInputBytes, maxOutputBytes]
      properties:
        enabled:
          type: boolean
        provider:
          type: string
        baseUrl:
          type: string
          description: 能力启用时必须是有效服务地址；禁用时允许为空字符串。
        model:
          type: string
          maxLength: 256
        quality:
          type: string
          enum: [auto, low, medium, high]
        timeout:
          type: string
          examples: [3m]
        maxInputBytes:
          type: integer
          minimum: 1
          maximum: 104857600
        maxOutputBytes:
          type: integer
          minimum: 1
          maximum: 104857600
    AISettingsSnapshot:
      type: object
      additionalProperties: false
      required: [text, imageModeration, imageToImage]
      properties:
        text:
          $ref: "#/components/schemas/TextAISettings"
        imageModeration:
          $ref: "#/components/schemas/ImageModerationAISettings"
        imageToImage:
          $ref: "#/components/schemas/ImageToImageAISettings"
    APIKeyStatus:
      type: object
      additionalProperties: false
      required: [configured, source]
      properties:
        configured:
          type: boolean
        hint:
          type: string
          description: 可选的脱敏末四位；永不包含完整 API Key。
        source:
          type: string
          enum: [environment, admin]
    APIKeyMutation:
      type: object
      additionalProperties: false
      required: [value, clear]
      properties:
        value:
          type: string
          maxLength: 4096
          writeOnly: true
          description: 留空表示保持现有值；该字段不会出现在任何响应中。
        clear:
          type: boolean
    APIKeyMutations:
      type: object
      additionalProperties: false
      required: [text, imageModeration, imageToImage]
      properties:
        text:
          $ref: "#/components/schemas/APIKeyMutation"
        imageModeration:
          $ref: "#/components/schemas/APIKeyMutation"
        imageToImage:
          $ref: "#/components/schemas/APIKeyMutation"
    AdminAISettings:
      type: object
      additionalProperties: false
      required: [dynamicEnabled, version, source, settings, apiKeys]
      properties:
        dynamicEnabled:
          type: boolean
        version:
          type: integer
          minimum: 0
        source:
          type: string
          enum: [environment, admin]
        settings:
          $ref: "#/components/schemas/AISettingsSnapshot"
        apiKeys:
          type: object
          additionalProperties: false
          required: [text, imageModeration, imageToImage]
          properties:
            text:
              $ref: "#/components/schemas/APIKeyStatus"
            imageModeration:
              $ref: "#/components/schemas/APIKeyStatus"
            imageToImage:
              $ref: "#/components/schemas/APIKeyStatus"
        updatedBy:
          type: string
        updatedAt:
          type: string
          format: date-time
    AdminAISettingsResponse:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/AdminAISettings"
        requestId:
          type: string
    UpdateAdminAISettingsRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, settings, apiKeys]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
        settings:
          $ref: "#/components/schemas/AISettingsSnapshot"
        apiKeys:
          $ref: "#/components/schemas/APIKeyMutations"
    TestAdminAIConnectionRequest:
      type: object
      additionalProperties: false
      required: [capability, settings, apiKeys]
      properties:
        capability:
          type: string
          enum: [text, image_moderation, image_to_image]
        settings:
          $ref: "#/components/schemas/AISettingsSnapshot"
        apiKeys:
          $ref: "#/components/schemas/APIKeyMutations"
    TestAdminAIConnectionResponse:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [capability, latencyMs]
          properties:
            capability:
              type: string
              enum: [text, image_moderation, image_to_image]
            latencyMs:
              type: integer
              minimum: 0
        requestId:
          type: string
    LoveLetterPolishRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
    LoveLetterPolishResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [polishedText, skipped]
          properties:
            polishedText:
              type: string
            skipped:
              type: boolean
              description: 未配置文本模型 Key、未调用外部模型时为 true
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
        requestId:
          type: string
...